refactor : fetch함수 로직 변경
