+++ src/pages/rolling/messageCreateForm/api.jsx
@@ -18,18 +18,21 @@ export function requsetProfileImgData() {
   return RequestAPI(BASE_URL_PATH, 'profile-images/');
 }
 
-// eslint-disable-next-line consistent-return
-export async function postCreateMessageData(id, messageInformation) {
-  try {
-    const postAPI = await fetch(`${BASE_URL_PATH}4-6/recipients/${id}/messages/`, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify(messageInformation),
-    });
-    return postAPI;
-  } catch (e) {
-    console.log(e);
+export function createMessage(id, messageInformation) {
+  const postAPI = fetch(`${BASE_URL_PATH}4-6/recipients/${id}/messages/`, {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'application/json',
+    },
+    body: JSON.stringify(messageInformation),
+  });
+  return postAPI;
+}
+
+export function errorHandling(boolean, status, callback) {
+  if (boolean) {
+    callback();
+  } else {
+    throw new Error(`요청 중 문제가 발생했습니다. 오류코드 ${status}`);
   }
 }
