+++ src/pages/rolling/messageCreateForm/components/TextEditor.jsx
@@ -1,13 +1,8 @@
 import React from 'react';
-// eslint-disable-next-line import/no-extraneous-dependencies
 import ReactQuill from 'react-quill';
-import styled from 'styled-components';
-// eslint-disable-next-line import/no-extraneous-dependencies
-import 'react-quill/dist/quill.snow.css';
 import './TextEditor.css';
+import 'react-quill/dist/quill.snow.css';
 import { useMessageFormContext } from '../context/MessageFormContext';
-
-const Box = styled.div``;
 
 const modules = {
   toolbar: [['bold', 'italic', 'underline'], [{ align: [] }, { color: [] }], ['clean']],
@@ -23,16 +18,7 @@ function TextEditor({ messageLength }) {
   };
 
   return (
-    <Box>
-      <ReactQuill
-        style={{ height: '250px' }}
-        theme="snow"
-        modules={modules}
-        formats={formats}
-        value={message || ''}
-        onChange={handleQuillChange}
-      />
-    </Box>
+    <ReactQuill theme="snow" modules={modules} formats={formats} value={message || ''} onChange={handleQuillChange} />
   );
 }
